third_party/blink/web_tests/fast/frames/sandboxed-iframe-scripting-02-expected.txt

CONSOLE MESSAGE: PASS: Executed script in data URL
Verify that sandboxed frames with sandbox='allow-scripts' can execute script from data: URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS disallowedExecuted is 0
PASS allowedExecuted is 1
PASS successfullyParsed is true

TEST COMPLETE
